Major updates to ecopathlite algorithms
In this update, I've finally updated my version of Ecopath so that it
should be able to handle all input models, including those with
multi-stanza groups.  Specific changes include:
- Full implementation of all 5 unknown-filling algorithms, including
  the generalized inverse.
- Correction of a few minor bugs in the way I converted between
  emigration & emigration rate and biomass accumulation & biomass
  accumulation rate.
- Correction of a few errors in Algorithm 3 (though I still have yet
  to find a model to test this one on).

I have found what seems to be an inconsistency in the way biomass
accumulation is handled in the algorithms.  All documentation suggests
that BA should be included with fisheries catch (Y) and net migration
(NM) in the export term used in all equations. However, it seems I can
only replicate the results from EwE6 if I drop this term (in effect,
assuming it is 0).  This seems like a pretty major bug to have persisted
this long in the EwE software, so I'll have to follow up there to see
if I'm misunderstanding something.

Also, a small bug correction was made to mdb2ewein.m so that it can
now read in tables with spaces in their name.