Changed file naming so all files are stored in same folder
Rather the previous method of referencing the javascript scripts
by their full path name, I'm now creating a temporary folder and
copying those scripts into that folder alongside the .json input
file and .html figure-rendering files.  This allows the html files
to be rendered in a browser with via a local server without having
to worry about how the relative pathnames relate to the localhost.