Streamlined size calculations for svg canvaz and margin
The previous code had a few inconsistencies with the way the
svg canvas was set up and the margin for the main g group
added.  Now, the renderfoodwebforce.m function returns these
sizes such that one can easily set up a figure and axis in
Matlab that matches these dimensions.