Created new interactive version
Adjusted foodweblayout.js and created a new index.html file that
creates a GUI interface to play around with parameters.  The
intent is that the user copies the following files to a working
folder:
 - index.html
 - foodweblayout.js
 - labeler.js
 - packages.js
 - seedrandom.min.js

Then a food web json file should be copied to the same folder,
renamed as foodweb.json.  Once the user is happy with the layout,
the SVG can be saved via the Chrome SVG Crowbar bookmarklet.  The
browser-specific solution isn't ideal, but it at least gets away
from the need for phantomJS.