Optimized expected trophic similarity calculation
The extpect trophic similarity between two groups is based on the
number of predator and prey groups each has.  Calculating the
hypergeometric distribution is the most time-consuming portion of
this calculation.  I rearranged the calculations so I calculate this
the minimum number of times possible (i.e. for each compo of #pred,