Added checkbox to mark completed user action
Added a checkbox that adds a DONE statement at the end of the
parameter list when checked.  This serves as a marker that the
user is done manipulating things.  It is used in foodweblayout.m
to signal the script to extract html and close the browser.

I'm sure something more elegant can be added down the line, but
this works for now.