Overhauled code to work with newer multi-level grouping
The old syntax for multilevel grouping was a bit confusing and
unwieldy.  I've now replaced it with a newer way of indicating
dendrogram-like grouping, based more closely on the output
of the trophicgroup.m function with hierarchical option.

Also updated the documentation to be a little clearer, as well
as consistent with the newer syntax.