Complete rewrite of layout algorithm for D3 version 4
The version 4 release of the D3 library included an overhaul to
the force simulation tool (formerly called the force layout) that
underlies the food web layout algorithm.  The new force simulation
provides much more user control over the forces acting on each node,
making it much more straightforward to prescribe the exact behavior
I originally wanted in this tool.

Built-in force functions now exist for things that I hackily included
in my old version, including collision avoidance and forcing to
prescribed y-position.

This version also allows for multiple link-like forces to be applied,
so in this version, the flux edges are used for positioning (with a
weak force) while the trophic grouping edges provide a stronger force.
The result is more repeatable, with far fewer nodes flying to the edges
of the canvas (as they did in the old version where flux edges exerted
no force), and therefore less manual intervention required from the
user.